Fix the hostname panic issue story [91881010]
The sliceOrEmptyVal function can return nil, but in mapToAppParams function
we are invoking the function sliceOrEmptyVal as shown below.
So on returning nil its panicking because accessing value from nil pointer by (* nil) which is wrong.
hostsArr := *sliceOrEmptyVal(yamlMap, hosts, &errs)

After fixing the output as

@cf5:~/go/src/github.com/cloudfoundry/cli$ ./out/cf push -f ~/cf-downloads/simple-go-web-app/manifest.yml
FAILED
Error reading manifest file:
Expected hosts to be a list of strings