Fix the random failure of manifest test cases
story #91289824
The test cases are failing randomly because of the map is used in the
removeDuplicatedValue  function which doesn't guaranteed the order.

Iteration order

When iterating over a map with a range loop, the iteration order is not specified and is not guaranteed to be the same from one iteration to the next. Since Go 1 the runtime randomizes map iteration order, as programmers relied on the stable iteration order of the previous implementation. If you require a stable iteration order you must maintain a separate data structure that specifies that order(https://blog.golang.org/go-maps-in-action).

I had written a sample app & called the removeDuplicatedValue & executed multipletimes the output
order is different in multiple executions